+++ src/components/Modal/ModalApproveAction/ModalApproveAction.jsx
@@ -1,38 +1,17 @@
-// import React, { useState } from 'react';
-// import axios from 'axios';
 import { useDispatch } from 'react-redux';
 import { deleteMyNotice } from 'redux/notices/notices-operations';
 
-// import Modal from '../Modal';
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever';
 import styles from './ModalApproveAction.module.scss';
 
-// import '../ModalApproveAction/';
-
 export const ModalApproveAction = ({ onClose, _id, title }) => {
-  // const [setShowModal] = useState(false);
-
-  // const closeModal = () => {
-  //   setShowModal(false);
-  // };
-
   const dispatch = useDispatch();
   const handleDelete = id => {
-    // async () => {
-    // try {
-    //   const response = await axios.delete('/api/advertisement');
-    //   console.log(response.data);
-
-    //   closeModal();
-    // } catch (error) {
-    //   console.error(error);
-    // }
     dispatch(deleteMyNotice(id));
     onClose();
   };
 
   return (
-    // <Modal onClose={onClose}>
     <div className={styles.wrapper}>
       <h1 className={styles.title}>Delete advertisement?</h1>
       <p className={styles.text}>
@@ -54,6 +33,5 @@ export const ModalApproveAction = ({ onClose, _id, title }) => {
         </button>
       </div>
     </div>
-    // </Modal>
   );
 };
