+++ src/components/NoticeCategoryItem/NoticeCategoryItem.module.scss
@@ -110,7 +110,7 @@
   gap: 4px;
   align-items: center;
   justify-content: center;
-  width: 80px;
+  min-width: 80px;
   height: 28px;
   padding: 2px 4px;
   border-radius: 16px;
