+++ src/redux/notices/notices-slice.js
@@ -1,11 +1,16 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import { getNotices, getNoticeById, getMyNotices, deleteMyNotice } from './notices-operations';
+import {
+  getNotices,
+  getNoticeById,
+  getMyNotices,
+  deleteMyNotice,
+  getFavoritesList,
+} from './notices-operations';
 
 const initialState = {
   items: [],
   item: {},
-  myItems: [],
   isLoading: false,
   error: null,
 };
@@ -43,16 +48,23 @@ export const noticesSlice = createSlice({
       .addCase(getMyNotices.fulfilled, (state, { payload }) => {
         state.isLoading = false;
         state.error = null;
-        state.myItems = payload;
+        state.items = payload;
       })
       .addCase(getMyNotices.rejected, handleRejected)
       .addCase(deleteMyNotice.pending, handlePending)
-      .addCase(deleteMyNotice.fulfilled, (state, {payload}) => {
+      .addCase(deleteMyNotice.fulfilled, (state, { payload }) => {
         state.isLoading = false;
         state.error = null;
         state.items = state.items.filter(({ _id }) => _id !== payload);
       })
-      .addCase(deleteMyNotice.rejected, handleRejected);
+      .addCase(deleteMyNotice.rejected, handleRejected)
+      .addCase(getFavoritesList.pending, handlePending)
+      .addCase(getFavoritesList.fulfilled, (state, { payload }) => {
+        state.loading = false;
+        state.error = null;
+        state.items = payload;
+      })
+      .addCase(getFavoritesList.rejected, handleRejected);
   },
 });
